Battery UI: make new battery page default, remove classic layout
