+++ i18n/ar.json
@@ -12,18 +12,6 @@
     "disconnect": "قطع الاتصال"
   },
   "batterySettings": {
-    "batteryLevel": "مستوى البطارية",
-    "bufferStart": {
-      "full": "عندما أوشكت على الامتلاء {soc}.",
-      "never": "فقط مع فائض كافٍ"
-    },
-    "legendBottomName": "أولوية المنزل",
-    "legendBottomSubline": "حتى تصل إلى {soc}.",
-    "legendMiddleName": "السيارة أولا",
-    "legendMiddleSubline": "عندما تكون بطارية المنزل فوق {soc}.",
-    "legendTopAutostart": "يبدأ تلقائيًا",
-    "legendTopName": "بطارية تدعم الشحن",
-    "legendTopSubline": "عندما تكون بطارية المنزل فوق {soc}.",
     "modalTitle": "إعدادات البطارية"
   },
   "config": {
