+++ i18n/bg.json
@@ -20,25 +20,7 @@
     "title": "Статус на оторизацията"
   },
   "batterySettings": {
-    "batteryLevel": "Батерия %",
-    "bufferStart": {
-      "above": "когато над {soc}.",
-      "full": "при {soc}.",
-      "never": "само с достатъчен излишък."
-    },
-    "capacity": "{energy} от {total}",
-    "control": "Управление на батерията",
-    "discharge": "Предотвратяване на разряд в бърз режим и планирано зареждане.",
-    "disclaimerHint": "Бележка:",
-    "disclaimerText": "Тези настройки засягат само соларния режим. Поведението при зареждане се коригира съответно.",
     "gridChargeTab": "Зареждане от мрежата",
-    "legendBottomName": "Приоритизиране на зареждането на домашната батерия",
-    "legendBottomSubline": "докато достигне {soc}.",
-    "legendMiddleName": "Приоритизиране на зареждането на превозното средство",
-    "legendMiddleSubline": "когато домашната батерия е над {soc}.",
-    "legendTopAutostart": "автоматично стартиране",
-    "legendTopName": "зареждане използвайки батерия",
-    "legendTopSubline": "когато домашната батерия е над {soc}.",
     "modalTitle": "Домашна батерия",
     "usageTab": "Използване на батерията"
   },
