+++ i18n/bs.json
@@ -20,25 +20,7 @@
     "title": "Status autorizacije"
   },
   "batterySettings": {
-    "batteryLevel": "Nivo baterije",
-    "bufferStart": {
-      "above": "kad je iznad {soc}.",
-      "full": "kad na {soc}.",
-      "never": "samo uz dovoljno viška."
-    },
-    "capacity": "{energy} od {total}",
-    "control": "Kontrola baterije",
-    "discharge": "Spriječite pražnjenje u brzom načinu rada i planirano punjenje.",
-    "disclaimerHint": "napomena:",
-    "disclaimerText": "Ova podešavanja utiču samo na solarni način rada. Ponašanje punjenja se u skladu s tim prilagođava.",
     "gridChargeTab": "Mrežno punjenje",
-    "legendBottomName": "Prioritetizirajte punjenje kućne baterije",
-    "legendBottomSubline": "dok ne dostigne {soc}.",
-    "legendMiddleName": "Prioritetizirajte punjenje vozila",
-    "legendMiddleSubline": "kada je kućna baterija iznad {soc}.",
-    "legendTopAutostart": "Pokreni automatski",
-    "legendTopName": "Punjenje vozila podržano baterijama",
-    "legendTopSubline": "kada je kućna baterija iznad {soc}.",
     "modalTitle": "Kućna baterija",
     "usageTab": "Upotreba baterije"
   },
