+++ i18n/cs.json
@@ -63,28 +63,7 @@
     }
   },
   "batterySettings": {
-    "batteryLevel": "Nabití baterie",
-    "bufferStart": {
-      "above": "když je nad {soc}.",
-      "full": "když je na {soc}.",
-      "never": "pouze když je dost přebytků."
-    },
-    "capacity": "{energy} z {total}",
-    "control": "Ovládání baterie",
-    "discharge": "Zabrání vybití baterie FVE v okamžitém režimu a během naplánovaného nabíjení.",
-    "disclaimerHint": "Poznámka:",
-    "disclaimerText": "Tato nastavení jsou platná pouze v solárním režimu. Chování dobíjení se upraví dle potřeby.",
     "gridChargeTab": "Nabíjení z distribuční sítě",
-    "legendBottomName": "Upřednostnit nabíjení baterie FVE",
-    "legendBottomSubline": "dokud nedosáhne hodnoty {soc}.",
-    "legendMiddleName": "Upřednostnit nabíjení vozidla",
-    "legendMiddleSubline": "když je baterie domu nad {soc}.",
-    "legendTopAutostart": "Spustí se automaticky",
-    "legendTopName": "Nabíjení vozidla s využitím baterie domu",
-    "legendTopSubline": "když je baterie domu nad {soc}.",
-    "legendTopSublineAbove": "když nad {soc}",
-    "legendTopSublineDisabled": "je {soc}.",
-    "legendTopSublineDisabledState": "zakázáno",
     "modalTitle": "Nastavení baterie",
     "noBattery": "Nejsou nakonfigurovány žádné baterie.",
     "usageTab": "Využití baterie"
