+++ i18n/el.json
@@ -20,25 +20,7 @@
     "title": "Κατάσταση εξουσιοδότησης"
   },
   "batterySettings": {
-    "batteryLevel": "Ποσοστό Μπαταρίας",
-    "bufferStart": {
-      "above": "όταν είναι πάνω από {soc}.",
-      "full": "όταν είναι {soc}.",
-      "never": "μόνο με αρκετό πλεόνασμα."
-    },
-    "capacity": "{energy} από {total}",
-    "control": "Έλεγχος Μπαταρίας",
-    "discharge": "Αποτροπή εκφόρτισης σε γρήγορη λειτουργία και προγραμματισμένη φόρτιση.",
-    "disclaimerHint": "Σημείωση:",
-    "disclaimerText": "Αυτές οι ρυθμίσεις επηρεάζουν μόνο την ηλιακή Φ/Β λειτουργία. Η συμπεριφορά φόρτισης προσαρμόζεται ανάλογα.",
     "gridChargeTab": "Φόρτιση δικτύου",
-    "legendBottomName": "Προτεραιότητα στη φόρτιση της οικιακής μπαταρίας",
-    "legendBottomSubline": "μέχρι να φτάσει {soc}.",
-    "legendMiddleName": "Προτεραιότητα στη φόρτιση οχήματος",
-    "legendMiddleSubline": "όταν η οικιακή μπαταρία είναι πάνω από {soc}.",
-    "legendTopAutostart": "Ξεκινά αυτόματα",
-    "legendTopName": "Φόρτιση οχήματος μέσω μπαταρίας",
-    "legendTopSubline": "όταν η οικιακή μπαταρία είναι πάνω από {soc}.",
     "modalTitle": "Οικιακή Μπαταρία",
     "usageTab": "Χρήση μπαταρίας"
   },
