+++ i18n/fi.json
@@ -63,28 +63,7 @@
     }
   },
   "batterySettings": {
-    "batteryLevel": "Akun varaustaso",
-    "bufferStart": {
-      "above": "kun ylittää {soc}.",
-      "full": "kun {soc}.",
-      "never": "ainoastaan kun riittävästi ylijäämää."
-    },
-    "capacity": "{energy} / {total}",
-    "control": "Akun hallinta",
-    "discharge": "Vältä purkamista nopeassa tilassa ja suunnitellussa latauksessa.",
-    "disclaimerHint": "Huomaa:",
-    "disclaimerText": "Nämä asetukset vaikuttavat vain aurinkotilaan. Latauskäyttäytymistä säädetään vastaavasti.",
     "gridChargeTab": "Lataus verkosta",
-    "legendBottomName": "Priorisoi kodin akun lataus",
-    "legendBottomSubline": "kunnes saavutetaan {soc}.",
-    "legendMiddleName": "Priorisoi ajoneuvon lataus",
-    "legendMiddleSubline": "kun kodin akku on yli {soc}.",
-    "legendTopAutostart": "Aloita automaattisesti",
-    "legendTopName": "Akulla tuettu ajoneuvon lataus",
-    "legendTopSubline": "kun kodin akku on yli {soc}.",
-    "legendTopSublineAbove": "kun yli {soc}",
-    "legendTopSublineDisabled": "on {soc}.",
-    "legendTopSublineDisabledState": "ei käytössä",
     "modalTitle": "Kodin akku",
     "noBattery": "Akkuja ei ole määritelty.",
     "usageTab": "Akun käyttö"
