+++ i18n/fr.json
@@ -63,28 +63,7 @@
     }
   },
   "batterySettings": {
-    "batteryLevel": "Niveau de la batterie",
-    "bufferStart": {
-      "above": "lorsque chargé à plus que {soc}.",
-      "full": "quand chargé à {soc}.",
-      "never": "seulement quand le surplus est suffisant."
-    },
-    "capacity": "{energy} sur {total}",
-    "control": "Contrôle batterie",
-    "discharge": "Empêcher la décharge en mode rapide et en charge planifiée.",
-    "disclaimerHint": "Note :",
-    "disclaimerText": "Ces réglages n'affectent que le mode solaire. Le comportement de charge est adapté en conséquence.",
     "gridChargeTab": "Chargement réseau",
-    "legendBottomName": "Prioriser le chargement de la batterie domestique",
-    "legendBottomSubline": "Jusqu'à ce que {soc} soit atteint.",
-    "legendMiddleName": "Prioriser le chargement du véhicule",
-    "legendMiddleSubline": "lorsque la batterie domestique est au-dessus de {soc}.",
-    "legendTopAutostart": "Démarrer automatiquement",
-    "legendTopName": "Charge véhicule par la batterie",
-    "legendTopSubline": "lorsque la batterie domestique est au-dessus de {soc}.",
-    "legendTopSublineAbove": "quand supérieur à {soc}",
-    "legendTopSublineDisabled": "est à {soc}.",
-    "legendTopSublineDisabledState": "désactivé",
     "modalTitle": "Batterie domestique",
     "noBattery": "Aucune batterie n'est configurée.",
     "usageTab": "Utilisation batterie"
