+++ i18n/hr.json
@@ -20,25 +20,7 @@
     "title": "Status autorizacije"
   },
   "batterySettings": {
-    "batteryLevel": "Razina baterije",
-    "bufferStart": {
-      "above": "kada je iznad {soc}.",
-      "full": "kada je na {soc}.",
-      "never": "samo s dovoljno viška."
-    },
-    "capacity": "{energy} od {total}",
-    "control": "Upravljanje baterijom",
-    "discharge": "Spriječite pražnjenje u brzom načinu rada i pri planiranom punjenju.",
-    "disclaimerHint": "Napomena:",
-    "disclaimerText": "Ove postavke utječu samo na solarni način rada. Ponašanje punjenja se tome prilagođava.",
     "gridChargeTab": "Punjenje iz mreže",
-    "legendBottomName": "Prioritet punjenja kućne baterije",
-    "legendBottomSubline": "dok ne dosegne {soc} napunjenosti.",
-    "legendMiddleName": "Prioritet punjenja vozila",
-    "legendMiddleSubline": "kada je kućna baterija iznad {soc}.",
-    "legendTopAutostart": "Automatsko pokretanje",
-    "legendTopName": "Punjenje vozila iz baterije",
-    "legendTopSubline": "kada je kućna baterija iznad {soc}.",
     "modalTitle": "Kućna baterija",
     "usageTab": "Upotreba baterije"
   },
