+++ i18n/lb.json
@@ -63,28 +63,7 @@
     }
   },
   "batterySettings": {
-    "batteryLevel": "Luedstand vun der Batterie",
-    "bufferStart": {
-      "above": "wann iwwer {soc}.",
-      "full": "wann op {soc}.",
-      "never": "nëmme mat genuch Iwwerschoss."
-    },
-    "capacity": "{energy} vun {total}",
-    "control": "Batterie steieren",
-    "discharge": "D'Entlueden am schnelle Modus a geplangte Lueden verhënneren.",
-    "disclaimerHint": "Notiz:",
-    "disclaimerText": "Dës Astellunge beaflossen nëmmen de Solarmodus. D'Luedeverhalen gëtt deementspriechend ugepasst.",
     "gridChargeTab": "Vum Netz oplueden",
-    "legendBottomName": "Prioritéit fir d'Oplueden vun der Hausbatterie",
-    "legendBottomSubline": "bis {soc} erreecht sinn.",
-    "legendMiddleName": "Prioritéit fir d'Oplueden vum Gefier",
-    "legendMiddleSubline": "wann d'Hausbatterie iwwer {soc} ass.",
-    "legendTopAutostart": "Automatesch starten",
-    "legendTopName": "Batterie-ënnerstëtzt Opluede vum Gefier",
-    "legendTopSubline": "wann d'Hausbatterie iwwer {soc} ass.",
-    "legendTopSublineAbove": "wann iwwert {soc}",
-    "legendTopSublineDisabled": "ass {soc}.",
-    "legendTopSublineDisabledState": "desaktivéiert",
     "modalTitle": "Hausbatterie",
     "noBattery": "Keng Batterien konfiguréiert.",
     "usageTab": "Batterieverbrauch"
