+++ i18n/lt.json
@@ -63,28 +63,7 @@
     }
   },
   "batterySettings": {
-    "batteryLevel": "Kaupiklio įkrova",
-    "bufferStart": {
-      "above": "kai virš {soc}.",
-      "full": "kai {soc}.",
-      "never": "tik su pakankamu pertekliumi."
-    },
-    "capacity": "{energy} iš {total}",
-    "control": "Kaupiklio valdymas",
-    "discharge": "Neleisti kaupiklio iškrovimo režime Greitas ir planiniame įkrovime.",
-    "disclaimerHint": "Pastaba:",
-    "disclaimerText": "Šie nustatymai paveikia tik įkrovimą Saulė. Įkrovimo algoritmas atitinkamai pakeičiamas.",
     "gridChargeTab": "Įkrovimas iš tinklo",
-    "legendBottomName": "Prioritetas namo kaupiklio įkrovimui",
-    "legendBottomSubline": "kol pasieks {soc}.",
-    "legendMiddleName": "Prioritetas automobilio įkrovimui",
-    "legendMiddleSubline": "kai namo kaupiklis yra virš {soc}.",
-    "legendTopAutostart": "Startuoti automatiškai",
-    "legendTopName": "Automobilio įkrovimas su kaupiklio pagalba",
-    "legendTopSubline": "kai kaupiklis yra virš {soc}.",
-    "legendTopSublineAbove": "kai virš {soc}",
-    "legendTopSublineDisabled": "yra {soc}.",
-    "legendTopSublineDisabledState": "išjungta",
     "modalTitle": "Namų kaupiklis",
     "noBattery": "Nėra sukonfigūruotų kaupiklių.",
     "usageTab": "Kaupiklio naudojimas"
