+++ i18n/ro.json
@@ -20,25 +20,7 @@
     "title": "Starea autorizației"
   },
   "batterySettings": {
-    "batteryLevel": "Nivelul bateriei",
-    "bufferStart": {
-      "above": "când deasupra {soc}.",
-      "full": "când atinge {soc}.",
-      "never": "numai cu un surplus suficient."
-    },
-    "capacity": "{energy} din {total}",
-    "control": "Controlul bateriei",
-    "discharge": "Preveniți descărcarea în modul rapid și încărcarea planificată.",
-    "disclaimerHint": "Nota:",
-    "disclaimerText": "Aceste setări afectează numai modul solar. Comportamentul de încărcare este ajustat în consecință.",
     "gridChargeTab": "Încărcare la rețea",
-    "legendBottomName": "prioritate casa",
-    "legendBottomSubline": "până când atinge {soc}.",
-    "legendMiddleName": "in primul rand vehiculul",
-    "legendMiddleSubline": "când bateria casei e peste {soc}.",
-    "legendTopAutostart": "porneste automat",
-    "legendTopName": "Incarcare ajutata de baterie",
-    "legendTopSubline": "când bateria casei e peste {soc}.",
     "modalTitle": "Setari bateriei",
     "usageTab": "Utilizarea bateriei"
   },
