+++ i18n/ru.json
@@ -63,28 +63,7 @@
     }
   },
   "batterySettings": {
-    "batteryLevel": "Уровень заряда аккумулятора",
-    "bufferStart": {
-      "above": "когда выше {soc}.",
-      "full": "когда на {soc}.",
-      "never": "только при наличии достаточного излишка."
-    },
-    "capacity": "{energy} из {total}",
-    "control": "Управление аккумулятором",
-    "discharge": "Предотвратить разряд в быстром режиме и запланированной зарядке.",
-    "disclaimerHint": "Примечание:",
-    "disclaimerText": "Эти настройки влияют только на режим солнечной батареи. Поведение зарядки изменяется соответствующе.",
     "gridChargeTab": "Зарядка сети",
-    "legendBottomName": "Приоритет зарядки домашнего аккумулятора",
-    "legendBottomSubline": "пока не достигнет {soc}.",
-    "legendMiddleName": "Приоритет зарядки электромобиля",
-    "legendMiddleSubline": "когда домашний аккумулятор заряжен выше {soc}.",
-    "legendTopAutostart": "Автоматический запуск",
-    "legendTopName": "Зарядка автомобиля с поддержкой от аккумулятора",
-    "legendTopSubline": "когда домашний аккумулятор заряжен выше {soc}.",
-    "legendTopSublineAbove": "когда заряд выше {soc}",
-    "legendTopSublineDisabled": "составляет {soc}.",
-    "legendTopSublineDisabledState": "неактивно",
     "modalTitle": "Домашний Аккумулятор",
     "noBattery": "Аккумуляторы не настроены.",
     "usageTab": "Использование аккумулятора"
