+++ i18n/sk.json
@@ -63,28 +63,7 @@
     }
   },
   "batterySettings": {
-    "batteryLevel": "Stav nabitia batérie",
-    "bufferStart": {
-      "above": "ak je nad {soc}.",
-      "full": "ak je {soc}.",
-      "never": "len s dostatočným prebytkom."
-    },
-    "capacity": "{energy} z {total}",
-    "control": "Ovládanie batérie",
-    "discharge": "Zabrániť vybitiu v rýchlom režime a plánovanom nabíjaní.",
-    "disclaimerHint": "Poznámka:",
-    "disclaimerText": "Tieto nastavenia ovplyvňujú iba solárny režim. Správa nabíjania je tomu prispôsoboná.",
     "gridChargeTab": "Nabíjanie zo siete",
-    "legendBottomName": "Uprednostniť nabíjanie domácej batérie",
-    "legendBottomSubline": "pokiaľ nedosiahne {soc}.",
-    "legendMiddleName": "Uprednostniť nabíjanie auta",
-    "legendMiddleSubline": "keď je domáca batéria nad {soc}.",
-    "legendTopAutostart": "Začať automaticky",
-    "legendTopName": "Batériou podporované nabíjanie auta",
-    "legendTopSubline": "keď je domáca batéria nad {soc}.",
-    "legendTopSublineAbove": "keď nad {soc}",
-    "legendTopSublineDisabled": "je {soc}.",
-    "legendTopSublineDisabledState": "vypnúť",
     "modalTitle": "Domáca batéria",
     "noBattery": "Žiadne baterky nakonfigurované.",
     "usageTab": "Využitie batérie"
